refactor(p-diff-sync): unify types via new perspective-diff-types crate
Introduce `crates/perspective-diff-types` as the single source of truth
for the wire types that both `perspective-diff-algorithm` and the
`perspective_diff_sync_integrity` zome need: PerspectiveDiff,
PerspectiveDiffEntryReference, Snapshot, HashReference,
LocalHashReference, HashBroadcast, LinkExpression, Triple,
ExpressionProof, plus the `Hash = HoloHash<Action>` alias and the
HasDiffParents / OpId traits.

The crate carries an optional `hdi` feature that decorates the entry
types with `SerializedBytes` derives and `app_entry!` calls. The
integrity zome enables it; the algorithm crate's standalone tests
build without it.

This eliminates two layers of duplication:
- `crates/perspective-diff-algorithm/src/diff_types.rs` (235 LOC)
  with parallel mirror types and a hand-rolled 39-byte Hash struct.
- `link_adapter/conversions.rs` (163 LOC) with link_to_algo /
  hash_to_algo / entry_ref_from_algo / etc.

All callsites in pull/render/revisions/snapshots/chunked_diffs and
both retriever impls now use the shared types directly — no field-
by-field copy, no per-hash byte shuffle.

Wake-23 Step 1 of the deeper PR-A extraction (per Nico's directive).
Step 2 (move pull/commit/render bodies into the algorithm crate) builds
on this foundation.

Verification:
- algo crate cargo test (--test-threads=1): 17/17 ✅
- perspective_diff_sync zome cargo test (--test-threads=1): 24/24 ✅
- rust-executor cargo test (--test-threads=1): 633/633 ✅
- integrity-zome WASM build: clean, 1.91 MB, no wasm-bindgen leakage
- DNA + happ build.sh: packages successfully
