+++ bootstrap-languages/p-diff-sync/hc-dna/zomes/perspective_diff_sync/src/link_adapter/tests.rs
@@ -1,10 +1,8 @@
-// Step 13b-C phase 2 (wake-15): these HDK-boundary tests retain their
-// ActionHash inputs (the original p-diff-sync wire types) but now feed
-// the algorithm-crate `Workspace` via the `hash_to_algo` conversion;
-// the trait bound switches from `PerspectiveDiffRetreiver` to
-// `algo::WorkspaceRetriever` (MockPerspectiveGraph implements both).
+// HDK-boundary tests that drive the algorithm-crate Workspace via the
+// MockPerspectiveGraph. With the shared `perspective-diff-types` crate,
+// hashes flow as `ActionHash` (= `algo::Hash`) directly — no conversion.
 //
-// The pure BFS coverage now lives in
+// The pure BFS coverage lives in
 // `perspective_diff_algorithm::workspace::tests`.
 
 #[test]
@@ -12,7 +10,6 @@ pub fn test_merge_fast_forward() {
     use hdk::prelude::*;
     use perspective_diff_algorithm as algo;
 
-    use crate::link_adapter::conversions::hash_to_algo;
     use crate::link_adapter::workspace::Workspace;
     use crate::retriever::{Associations, GraphInput, MockPerspectiveGraph, GLOBAL_MOCKED_GRAPH};
 
@@ -48,14 +45,11 @@ pub fn test_merge_fast_forward() {
 
     let mut workspace = Workspace::new();
     let res = workspace.collect_until_common_ancestor::<MockPerspectiveGraph>(
-        hash_to_algo(&ActionHash::from_raw_36(vec![5; 36])),
-        hash_to_algo(&ActionHash::from_raw_36(vec![4; 36])),
+        ActionHash::from_raw_36(vec![5; 36]),
+        ActionHash::from_raw_36(vec![4; 36]),
     );
     assert!(res.is_ok());
-    assert_eq!(
-        res.unwrap(),
-        hash_to_algo(&ActionHash::from_raw_36(vec![0; 36]))
-    );
+    assert_eq!(res.unwrap(), ActionHash::from_raw_36(vec![0; 36]));
     let _ = algo::null_node;
 }
 
@@ -64,7 +58,6 @@ pub fn test_fork_with_none_source() {
     use hdk::prelude::*;
     use perspective_diff_algorithm as algo;
 
-    use crate::link_adapter::conversions::hash_to_algo;
     use crate::link_adapter::workspace::Workspace;
     use crate::retriever::{GraphInput, MockPerspectiveGraph, GLOBAL_MOCKED_GRAPH};
 
@@ -79,8 +72,8 @@ pub fn test_fork_with_none_source() {
 
     let mut workspace = Workspace::new();
     let res = workspace.collect_until_common_ancestor::<MockPerspectiveGraph>(
-        hash_to_algo(&ActionHash::from_raw_36(vec![0; 36])),
-        hash_to_algo(&ActionHash::from_raw_36(vec![1; 36])),
+        ActionHash::from_raw_36(vec![0; 36]),
+        ActionHash::from_raw_36(vec![1; 36]),
     );
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), algo::null_node());
@@ -91,7 +84,6 @@ pub fn test_merge_fast_forward_none_source() {
     use hdk::prelude::*;
     use perspective_diff_algorithm as algo;
 
-    use crate::link_adapter::conversions::hash_to_algo;
     use crate::link_adapter::workspace::Workspace;
     use crate::retriever::{Associations, GraphInput, MockPerspectiveGraph, GLOBAL_MOCKED_GRAPH};
 
@@ -109,8 +101,8 @@ pub fn test_merge_fast_forward_none_source() {
 
     let mut workspace = Workspace::new();
     let res = workspace.collect_until_common_ancestor::<MockPerspectiveGraph>(
-        hash_to_algo(&ActionHash::from_raw_36(vec![2; 36])),
-        hash_to_algo(&ActionHash::from_raw_36(vec![1; 36])),
+        ActionHash::from_raw_36(vec![2; 36]),
+        ActionHash::from_raw_36(vec![1; 36]),
     );
     assert!(res.is_ok());
     assert_eq!(res.unwrap(), algo::null_node());
