+++ bootstrap-languages/p-diff-sync/hc-dna/zomes/perspective_diff_sync_integrity/src/lib.rs
@@ -1,118 +1,20 @@
 use chrono::{DateTime, Utc};
 use hdi::prelude::*;
 
-// Bridge to the substrate-agnostic algorithm crate (Step 1.5):
-// implementing `HasDiffParents` on `PerspectiveDiffEntryReference` here —
-// rather than in p-diff-sync — sidesteps the orphan rule, since both
-// the trait (in `perspective_diff_algorithm`) and the type (here in
-// `perspective_diff_sync_integrity`) are foreign to p-diff-sync.
-use perspective_diff_algorithm::HasDiffParents;
-
-#[derive(
-    Serialize, Deserialize, Clone, SerializedBytes, Debug, PartialEq, Eq, Hash, Ord, PartialOrd,
-)]
-pub struct ExpressionProof {
-    pub signature: String,
-    pub key: String,
-}
-
-#[derive(
-    Serialize, Deserialize, Clone, SerializedBytes, Debug, PartialEq, Eq, Hash, Ord, PartialOrd,
-)]
-pub struct Triple {
-    pub source: Option<String>,
-    pub target: Option<String>,
-    pub predicate: Option<String>,
-}
-
-#[derive(Clone, Deserialize, Serialize, Debug, PartialEq, Eq, Hash, Ord, PartialOrd)]
-pub struct LinkExpression {
-    pub author: String,
-    pub data: Triple,
-    pub timestamp: String,
-    pub proof: ExpressionProof,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes, Default, PartialEq, Eq, Hash, Ord, PartialOrd)]
-pub struct PerspectiveDiff {
-    pub additions: Vec<LinkExpression>,
-    pub removals: Vec<LinkExpression>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes)]
-pub struct CommitInput {
-    pub diff: PerspectiveDiff,
-    pub my_did: String,
-}
-
-///The reference that is sent to other agents, denotes the position in the DAG as well as the data at that position
-#[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes)]
-pub struct HashBroadcast {
-    pub reference_hash: HoloHash<holo_hash::hash_type::Action>,
-    pub reference: PerspectiveDiffEntryReference,
-    pub broadcast_author: String,
-}
-
-impl PerspectiveDiff {
-    pub fn new() -> Self {
-        Self {
-            additions: Vec::new(),
-            removals: Vec::new(),
-        }
-    }
-    pub fn total_diff_number(&self) -> usize {
-        self.additions.len() + self.removals.len()
-    }
-    
-    pub fn get_sb(self) -> ExternResult<SerializedBytes> {
-        self.try_into()
-            .map_err(|error| wasm_error!(WasmErrorInner::Host(String::from(error))))
-    }
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes)]
-pub struct Snapshot {
-    pub diff_chunks: Vec<HoloHash<holo_hash::hash_type::Action>>,
-    pub included_diffs: Vec<HoloHash<holo_hash::hash_type::Action>>,
-}
-
-app_entry!(Snapshot);
-
-#[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes, PartialEq, Eq, Hash)]
-pub struct PerspectiveDiffEntryReference {
-    pub diff: PerspectiveDiff,
-    pub parents: Option<Vec<HoloHash<holo_hash::hash_type::Action>>>,
-    pub diffs_since_snapshot: usize,
-    /// Optional hashes of chunked diff entries for large diffs.
-    /// When this is Some and non-empty, the `diff` field should be empty/default
-    /// and the actual diff data is stored in separate chunk entries.
-    #[serde(default)]
-    pub diff_chunks: Option<Vec<HoloHash<holo_hash::hash_type::Action>>>,
-}
-
-app_entry!(PerspectiveDiffEntryReference);
+// The shared wire types live in `perspective-diff-types`; this crate
+// re-exports them so the rest of p-diff-sync can keep importing from
+// `perspective_diff_sync_integrity::*` while the algorithm crate
+// consumes the same struct shapes via `perspective_diff_types`.
+pub use perspective_diff_types::{
+    null_node, CommitInput, ExpressionProof, HasDiffParents, HashBroadcast, HashReference,
+    LinkExpression, LocalHashReference, OpId, PerspectiveDiff, PerspectiveDiffEntryReference,
+    PullResult, Snapshot, Triple,
+};
 
 #[derive(Clone, Serialize, Deserialize, Debug)]
 pub struct Perspective {
     pub links: Vec<LinkExpression>,
 }
-
-//TODO: this can likely be removed and instead just reference the PerspectiveDiffEntry/MergeEntry directly?
-#[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes)]
-pub struct HashReference {
-    pub hash: HoloHash<holo_hash::hash_type::Action>,
-    pub timestamp: DateTime<Utc>,
-}
-
-app_entry!(HashReference);
-
-#[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes)]
-pub struct LocalHashReference {
-    pub hash: HoloHash<holo_hash::hash_type::Action>,
-    pub timestamp: DateTime<Utc>,
-}
-
-app_entry!(LocalHashReference);
 
 #[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes)]
 pub struct LocalTimestampReference {
@@ -136,8 +38,8 @@ pub struct PerspectiveExpression {
 
 app_entry!(PerspectiveExpression);
 
-/// Signal payload that includes recipient DID for multi-user routing
-/// Flattened structure to avoid Holochain extracting nested PerspectiveExpression
+/// Signal payload that includes recipient DID for multi-user routing.
+/// Flattened structure to avoid Holochain extracting nested PerspectiveExpression.
 #[derive(Clone, Debug, Serialize, Deserialize, SerializedBytes)]
 pub struct RoutedSignalPayload {
     pub recipient_did: String,
@@ -158,12 +60,6 @@ pub struct OnlineAgentAndAction {
     pub did: String,
     pub status: Option<PerspectiveExpression>,
     pub status_action: Option<ActionHash>,
-}
-
-#[derive(Debug, Serialize, Deserialize, SerializedBytes)]
-pub struct PullResult {
-    pub diff: PerspectiveDiff,
-    pub current_revision: Option<HoloHash<holo_hash::hash_type::Action>>,
 }
 
 #[hdk_entry_types]
@@ -216,78 +112,6 @@ impl RoutedSignalPayload {
     }
 }
 
-impl HashBroadcast {
-    pub fn get_sb(self) -> ExternResult<SerializedBytes> {
-        self.try_into()
-            .map_err(|error| wasm_error!(WasmErrorInner::Host(String::from(error))))
-    }
-}
-
-impl PerspectiveDiffEntryReference {
-    pub fn new(
-        diff: PerspectiveDiff,
-        parents: Option<Vec<HoloHash<holo_hash::hash_type::Action>>>,
-    ) -> Self {
-        Self {
-            diff: diff,
-            parents: parents,
-            diffs_since_snapshot: 0,
-            diff_chunks: None,
-        }
-    }
-
-    /// Create a new entry reference with chunked diffs
-    pub fn new_chunked(
-        diff_chunks: Vec<HoloHash<holo_hash::hash_type::Action>>,
-        parents: Option<Vec<HoloHash<holo_hash::hash_type::Action>>>,
-        diffs_since_snapshot: usize,
-    ) -> Self {
-        Self {
-            diff: PerspectiveDiff::new(), // Empty diff when using chunks
-            parents: parents,
-            diffs_since_snapshot: diffs_since_snapshot,
-            diff_chunks: Some(diff_chunks),
-        }
-    }
-
-    /// Check if this entry uses chunked storage
-    pub fn is_chunked(&self) -> bool {
-        self.diff_chunks.as_ref().map_or(false, |chunks| !chunks.is_empty())
-    }
-
-    /// Backward compatibility method to extract the diff data
-    pub fn to_perspective_diff(&self) -> PerspectiveDiff {
-        self.diff.clone()
-    }
-
-    /// Helper method to get the comparison key for ordering
-    // Compare using tuple ordering: entries with parents come first,
-    // then by parent hashes, then by diffs_since_snapshot,
-    // then by total diff count, then by diff contents
-    fn comparison_key(&self) -> (bool, &Option<Vec<HoloHash<holo_hash::hash_type::Action>>>, usize, usize, &PerspectiveDiff) {
-        let has_parents = self.parents.is_some();
-        (!has_parents, &self.parents, self.diffs_since_snapshot, self.diff.total_diff_number(), &self.diff)
-    }
-}
-
-impl PartialOrd for PerspectiveDiffEntryReference {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for PerspectiveDiffEntryReference {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        self.comparison_key().cmp(&other.comparison_key())
-    }
-}
-
-impl HasDiffParents<HoloHash<holo_hash::hash_type::Action>> for PerspectiveDiffEntryReference {
-    fn parents(&self) -> Option<&[HoloHash<holo_hash::hash_type::Action>]> {
-        self.parents.as_deref()
-    }
-}
-
 impl OnlineAgent {
     pub fn get_sb(self) -> ExternResult<SerializedBytes> {
         self.try_into()
@@ -310,19 +134,16 @@ pub fn validate(op: Op) -> ExternResult<ValidateCallbackResult> {
                 // Validate parent dependencies
                 if let Some(parents) = pdiff_ref.parents {
                     for parent_action_hash in parents {
-                        // Ensure each declared parent exists and is valid in the source chain/DHT
                         if must_get_valid_record(parent_action_hash.clone()).is_err() {
                             missing.push(parent_action_hash.into());
                         }
                     }
                 }
 
-                // Validate chunk dependencies
-                // Chunks must be available before the parent entry can be validated.
-                // The commit flow ensures chunks are created and validated BEFORE the parent.
+                // Validate chunk dependencies — chunks must be available before
+                // the parent entry can be validated.
                 if let Some(diff_chunks) = pdiff_ref.diff_chunks {
                     for chunk_action_hash in diff_chunks {
-                        // Ensure each chunk entry exists and is valid
                         if must_get_valid_record(chunk_action_hash.clone()).is_err() {
                             missing.push(chunk_action_hash.into());
                         }
@@ -330,13 +151,14 @@ pub fn validate(op: Op) -> ExternResult<ValidateCallbackResult> {
                 }
 
                 if !missing.is_empty() {
-                    return Ok(ValidateCallbackResult::UnresolvedDependencies(UnresolvedDependencies::Hashes(missing)));
+                    return Ok(ValidateCallbackResult::UnresolvedDependencies(
+                        UnresolvedDependencies::Hashes(missing),
+                    ));
                 }
             }
 
             Ok(ValidateCallbackResult::Valid)
         }
-        // For all other ops in this zome we accept by default
         _ => Ok(ValidateCallbackResult::Valid),
     }
 }
