+++ crates/perspective-diff-algorithm/Cargo.toml
@@ -20,6 +20,7 @@ chrono = { version = "0.4", default-features = false, features = ["clock", "std"
 serde = { version = "1", features = ["derive"] }
 thiserror = "1"
 petgraph = "0.6"
+perspective-diff-types = { path = "../perspective-diff-types" }
 
 [dev-dependencies]
 once_cell = "1"
